# Add redis-rb with hiredis to the perf benchmarks