@davidor 2019-03-04
I noticed that Travis also runs the tests on TruffleRuby and JRuby those fail because they do not support C extensions.

I'll try to provide a fix for that tomorrow.