@davidor 2019-03-05
I think that the issue should be fixed.

I see that JRuby tests are still failing but it's because of #13 . TruffleRuby is failing as well, but it looks like it's not because of something introduced in this PR, because my other PR fails too in TruffleRuby #14 with the same error.